Remodularisation of the code - should be no functional change however extensive testing needed
