+++ Tools/HILSIM_XPlane/Display.h
@@ -1,0 +1,7 @@
+#pragma once
+
+void DisplayStart(void);
+void DisplayStop(void);
+void DisplayEnable(void);
+void DisplayDisable(void);
+
