+++ Tools/HILSIM_XPlane/Servos.h
@@ -1,0 +1,8 @@
+#pragma once
+
+void SetupDefaultServoZeros(void);
+void ServosToControls(void);
+void ServosEnable(string& OverString);
+float ServosBrakeSetting(void);
+float* ServosThrottleSettings(void);
+
