+++ Tools/HILSIM_XPlane/UDBSocketUnix.c
@@ -62,7 +62,7 @@ void SetTermIOs(void)
 	tcsetattr(STDIN_FILENO, TCSANOW, &ttystate);
 }
 
-UDBSocket UDBSocket_init(UDBSocketType type, uint16_t UDP_port, char* UDP_host, char* serial_port, long serial_baud)
+UDBSocket UDBSocket_init(UDBSocketType type, uint16_t UDP_port, const char* UDP_host, const char* serial_port, long serial_baud)
 {
 #if (NIX == 1)
 	speed_t BAUD;
@@ -153,6 +153,10 @@ UDBSocket UDBSocket_init(UDBSocketType type, uint16_t UDP_port, char* UDP_host, 
 		{
 			struct termios  config;
 
+			if (NULL == newSocket->serial_port) {
+				free(newSocket);
+				return NULL;
+			}
 			newSocket->fd = open(newSocket->serial_port, O_RDWR | O_NOCTTY | O_NDELAY);
 			if (newSocket->fd == -1) {
 				perror("serial open() failed");
@@ -318,7 +322,7 @@ void UDBSocket_close(UDBSocket socket)
 		case UDBSocketUDPServer:
 		{
 			shutdown(socket->fd, SHUT_RDWR);
-			close(socket->fd);
+			if (socket->fd >= 0) close(socket->fd);
 			if (socket->UDP_host) free(socket->UDP_host);
 			if (socket->serial_port) free(socket->serial_port);
 			free(socket);
@@ -326,7 +330,7 @@ void UDBSocket_close(UDBSocket socket)
 		}
 		case UDBSocketSerial:
 		{
-			close(socket->fd);
+			if (socket->fd >= 0) close(socket->fd);
 			if (socket->serial_port) free(socket->serial_port);
 			free(socket);
 			break;
@@ -355,7 +359,10 @@ int UDBSocket_read(UDBSocket socket, unsigned char* buffer, int bufferLength)
 				select(STDIN_FILENO+1, &fds, NULL, NULL, &tv);
 				if (FD_ISSET(STDIN_FILENO, &fds)) // only read stdin if there's data ready
 				{
-					buffer[pos++] = fgetc(stdin);
+					int c;
+					c = fgetc(stdin);
+					if (EOF == c) break;
+					buffer[pos++] = (unsigned char)c;
 				}
 				else
 				{
@@ -408,7 +415,7 @@ int UDBSocket_read(UDBSocket socket, unsigned char* buffer, int bufferLength)
 	return -1;
 }
 
-int UDBSocket_write(UDBSocket socket, unsigned char* data, int dataLength)
+int UDBSocket_write(UDBSocket socket, const unsigned char* data, int dataLength)
 {
 	switch (socket->type)
 	{
